Alterações na terms_page
