--- lib/screens/service/components/question_text.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class QuestionText extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        padding: EdgeInsets.only(top: 70, left: 20, right: 20),
-        child: Text(
-          'Você recebeu a ajuda que necessitava?',
-          style: TextStyle(
-              color: Color(0xFF2a2a2a),
-              fontSize: 18,
-              fontWeight: FontWeight.bold),
-        ));
-  }
-}
