+++ lib/screens/terms/components/default_subtitle.dart
@@ -8,13 +8,14 @@ class DefaultSubtitle extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-        padding: EdgeInsets.only(left: 20, right: 20, bottom: 10),
-        child: Text(
-          text,
-          style: TextStyle(
-              color: Color(0xFF474747),
-              fontSize: 25,
-              fontWeight: FontWeight.bold),
-        ));
+      padding: EdgeInsets.only(bottom: 10),
+      child: Text(
+        text,
+        style: TextStyle(
+            color: Color(0xFF474747),
+            fontSize: 25,
+            fontWeight: FontWeight.bold),
+      ),
+    );
   }
 }
